BUGFIX: Use preview links consistently
We introduced the redirect links for performance in
older Neos versions to prevent resolving routes for
Nodes which was quite expensive. The Neos 9 preview
links are easy to resolve and we have no need for
the redirect links for previews. This helps keep
preview links consistent and always use the previewAction.