TASK: Introduce aliased `useSelector` hook for Neos global State
This can be used in functional components to avoid having to use the `connect()` decorator or hoc

While a direct import from "react-redux" would work, the `useSelector` would not be aware of its state type. And as the defaultState was removed https://github.com/reduxjs/react-redux/issues/1879 we resort to now cleanly reexporting it.