WIP: LinkEditor simplify Web dialog by allowing all kinds of links and protocols and also include the `anchor` option as plain text
this allows insert urls like

```
seb://bist-du-glücklich?ja=true#test
```

Instead of hard validation errors we now emit warnings as in that the apply button is not blocked.

To allow determining the anchor we disable the anchor text field option as there is no two-way-binding in the LinkType framework yet and its questionable if its sensible.