!!! FEATURE: Simplify link editor by not separating anchor from href and displaying it always separately
With the configuration `anchor: true` an editor option for the anchor was previously enabled. This was necessary in the old link editor to enable anchors on node urls. The old link editor though did allow inserting an anchor into the manual link field either way - enabling the anchor field would allow to edit this segment separately with a two-way binding. In the archaeopteryx the pattern of having to enable anchors was continued and the web tab forced to specify a base url and did not allow relative anchors until recently.

A recent hack now allowed to edit the full url with anchor in the web tab by returning in `convertModelToLink` also the `options.anchor` just to have that in the inspector glued back together again. This clearly breaks the abstraction, and it became imminent that the anchor is part of the href and should not be treated as separate option.

The anchor option is now displayed separately for the document and asset tab. The web tab will allow the anchor to be set in the main field together with the rest of the link.

Breaking changes:
----------------

The global editor option for the inspector and for the inline link editor `anchor: <boolean>` is no longer respected.

Instead, anchors are enabled for the asset and node tab at all times.