TASK: Revert introducing big library `libphonenumber-js` for phone number validation and country calling code suggestions
The library is quite big due to the dataset. Now while it gives some validation when parsing a phone number and also can help by finding the correct country calling code - these features do not justify the complexity of the new library.

We expect that phone numbers are usually inserted as is with or without country code. Its seldom of need for a neos editor to have an overview within neos of all 300+ country calling codes and its not trivial to build good ux for that.

For further simplification phone numbers are now a simple text field with basic warnings if empty or non digits are entered.