FEATURE: Position Dialog Modals centered and adjust width
The dialogs always leave a spacing of 20vh from the top which wastes spaces and looks unbalanced on laptops as below the dialog ends often without gap.

This changes centers the dialogs now on the vertical axis so they are always in center no matter their height. (see small vs. big creation dialog)

Additionally, the default width of "wide" was a little reduced as select boxes and text-fields in the creation dialog grew super wide - much wider than one is used from them in the inspector. This size adjustment will also make the new link editor appear by default in the correct width.

The dialog additionally grows by default and in "wide" mode to fit its contents. This is required when the new link editor shows the asset selection which takes the full width.

For reference the original Archaeopterix employed a hack to archive this which is now merged into the core behaviour:

https://github.com/sitegeist/Sitegeist.Archaeopteryx/blob/2d53b63b483b608dcc062a23bf2af608f27d5445/Neos.Ui/core/src/presentation/Modal.tsx#L7-L34