TASK: Introduce api to use global state for global registry
Originally the Neos Ui was build seemingly without real global state. There is a high order component function - invoked via a decorator @neos() which allowed to inject the registry.

For functional components currently one would need to use the same high-order-pattern which is a little quirky with functions compared to say the new concept of hooks.

The use of hooks as approach was shown in `@neos-project/neos-ui-link-editor-neos-bridge` we would continue to use the NeosContext and fetch the global registry or configuration from there via hooks like

- useNeos()
- useGlobalRegistry()
- ...useConfiguration()

but as this information is truly global and never leads to a re-rendering e.g. is static. We dont have to make that look like a hook but just use javascript state instead. (Like we do for the new translate() function)