TASK: Separate state from configuration -> move the `allowedTargetWorkspaces` into the redux store
previously these options where fully static and part of the "configuration".
But this is misleading, the information depends on other workspaces and thus the database. It is not configuration.

Now they will also be updated when `const {getWorkspaceInfo} = backend.get().endpoints;` is invoked.

During changes only the totalNumberOfChanges are recalculated but not all the allowedTargetWorkspaces