FEATURE: Ensure plugin api works by providing e2e test
In the test distribution this change adds a test plugin wich is also registered as yarn workspace so we can use esbuild and more importantly the neos-ui-extensibility to test.

via

```
yarn workspace @neos-project/neos-ui-test-plugin run build
```

this plugin is build and the esbuild.js script ensures with a custom analysis that we only bundle files from the source and from the extensibility shims but not any other node_modules

An e2e test run via testcafe ensures that the plugin is correctly registered via yaml and initialised. A client function access values from the plugin that were exposed on the window.