BUGFIX: Prevent empty content tree when incomplete node data for pagetree is merged (#4024)
With #3961 we don’t return content type childnodes in not „fully loaded“ node data for the ui anymore.
When the pagetree data is merged after the content tree data (which is fully loaded) this lead to
an empty page tree as the childnodes in the store were overriden.

With this change we ignore a partial childnode list when merging if we already have a „fully loaded“ node in the store.

Resolves: #4023