TASK: Remove now unused `@neos-project/utils-helpers` `isEmail` helper in favour of `isemail`
the library was introduced with https://github.com/neos/neos-ui/pull/2988 and handles more cases. A regex would probably be faster and less bloat but as we have the lib already.

For the MailTo field a simple stupid validator like `a-z@a-z.a-z` would probably do already as the worst thing is that a warning is emitted - we dont block the change. And if some redactor enters an email and forgets the tld its probably a mistake and one is not running on a local server with `mail@server1` - but as said it removed things for now so :)