FEATURE: Group href and link options in "advanced options" details below
This is the third approach to where the options could reside. Initially they were below with anchor and title and rel mixed. Then we - correctly - separated anchors from the options as in `Web` we want to allow to insert a simple anchor directly. And we moved the other link options to the top next to the preview as we wanted to symbolize that they are really available everywhere for every link - and that they affect the set link and are not dependent on the tab. But in the dresden sprint it was found that the anchor as well as the others are really "Advanced" in some common way so we want to group and hide them together below in every tab.