TASK: Simplify Link Editor framework by removing first class support for asynchronous models
There is a lot of complexity to async  `convertLinkToModel` because we cannot initially merge this value into the linkTypes observeable state -> this was not a problem as all `Editor` work with a nullable `State` either way as its would also be null if not loaded.

But we also want to open the additional link options if the link has `anchor` set for example. For that we need the resolved link model see linktype.isValid and we were only be able to compute this one tick later. That creates glitches and content shifts for the opening animation of the dialog.

Technically we could solve that by strict separating `DialogWithValue` into `DialogWithValue` and `DialogWithAsyncValue` where the latter would receive a promise and use `usePromise` to resolve that. Splitting it this way would allow non async values to appear immediately which is the default case and async values to be displayed after a loading animation.

Now as the core does not need any async link editors we dont need to add this complexity right now just to carry it around. In case a plugin wants to register a dialog with async state it would need to simplify manage that itself.