TASK: Use `#[Flow\Route]` instead of unused `'@action': 'ignored'` and deprecated `@subpackage`
This hopefully makes christian happy.

We place the controllers to handle the endpoints into the `LinkEditor` namespace. To archive that on normally has to use `@subpackage` to descend into the namespace. We dont want to promote that behaviour and instead want to make controllers discoverable in all places and identify them simply by FQN.
Also for these query endpoints we currently didnt use the ActionController but simply the `ControllerInterface` and handled the `ActionRequest` to a single `processQuery`. This means we ignored the `@action` route value. To make that transparent we wrote `@ignored` into that - by default it would just be `index`. This also seems we work more around the framework and we dont want to look like we do :) -> in the future flow should provide a way to legally omit the `@action` and have it default to `processRequest` or something.

As these things are not implemented in Flow we make use of the `#[Flow\Route]` attribute which already as part of forwards compatibility allows to place controllers quite flexible (it will still use the @subpackage bts) but its part of the core and allowed to be used this way.
Also we now use the @action routing again but we just simply by convention only use one per controller.

The `AbstractQueryController` behaves now like and `ActionController` but doesnt invoke all the crazy property mapping and things we dont want to use :)