TASK: Prefer conventional jest tests over not run native node tests
The `Sitegeist.Archaeopteryx` went with https://github.com/sitegeist/Sitegeist.Archaeopteryx/pull/83 in the direction to abandon jest. We also might plan that for Neos as jest was seemingly dead and brings a lot of ugly dev dependencies like babel.
This link editor change also introduced this new node test running but it was commented out as the new dependencies need to be organised first. That leaves us with not run `test.js` files. Now we reevaluated that we probably have to keep jest alive for some time still and thus all the new tests are now adjusted to jest style.