TASK: Overhaul dialog title and apply button labels
Instead of a varying label for Create {linkType} Link - which is hard to translate for example in german and still looks odd if quoted "Dokument"-Link erstellen - we have now a static title based on if there was an initial link or not.

We also change the label of the button if a link is updated or newly created - this should allow a more clear workflow as even using a new tab we still update a link and not as before create a new one. (Same with delete)