TASK: Simplify link editor dialog rendering by combining `DialogWithEmptyValue` with `DialogWithValue`
And also get rid of hacky details like `initialLinkTypePreview?: () => React.ReactNode`

Initially the split into two components and the duplication of the tab rendering and the like was necessary as the initial link type allowed upon `convertLinkToModel` to invoke a react hook. And react hooks can not be used conditionally.

 With this removal of first class async tabs (a3132ca5cc96426d06fe37e8fd8d00d696e1ba71) we can also simplify the code in the Dialog.