BUGFIX: Dont use `translate(-50%, -50%)` to position dialog as this trips off popover positioning
It seems the transformation is technically part of the spec as per this discussion: https://github.com/w3c/csswg-drafts/issues/8584
But not implemented in any browser.
The transformation is currently not taken into account and the popover is positioned as if the dialog is not transformed.

To fix that we now resort to a simple flex centering.

TODO: Due to flex and no longer the use of `left: 50%;` the width of the dialog is not sanely restricted to 50vw but always as wide as possible making text dialogs pretty wide.