BUGFIX: CKEditor focus tracker must not use host frame `document`
We cannot access "document" as this will always be the host frame.
If ckeditor is used in the guest iframe we need to use their document.

The major problem here is the `_blur` method.
Technically relying on the first "ownerDocument" of the possibly multiple elements is not correct as they _could_ belong to different documents but in practice they dont and should not.

Without that patch in the `_blur` method the buttons in the toolbar of the ckeditor could not be toggled correctly. Instead of getting focus on the button on pressing the button and closing the dropdown on release, the focus already lead to closing the dropdown and the mouse release - if on that button - would reopen it again leading to a buggy behaviour. Any dropdown like the paragraph dropdown or to show more options was affected.