BUGFIX: CKEditor tooltip manager must not track mouse in host frame `document`
We need to track the mouse position IN the iframe where the CKEditor is mounted as the document itself is the host frame.
This patch fully opts in to only support the iframe. CKEditor mounted in the host frame - for example the RTE Editor will not show tooltips.
That is the bargain. To fix both cases we need to actually boot and load CKEditor in the iframe by using code sharing.

We cannot support both cases via a simple patch as TooltipManager uses as singleton pattern. We would need to change multiple parts and logic.