FEATURE: Show outline while editing text
CKEditor natively adds the `.ck-focused` class to active editables.
It also styles that by using a greyish css border.
With the new inline overhaul we load this style but it does not fit with the known dotted blue outline which is visible beforehand on hovering. The ckeditor border is positioned tighter and also due to the difference in color the behaviour when clicking into editables feels wrong.

Now instead of fully disabling the border we want to introduce one ourselves as its a missing feature

see https://github.com/neos/neos-ui/issues/3853

The dottet border known during hover will now be kept during editing a node. There is one exception: When the node only has one html element and that is the ckeditor editable we cannot show two outlines as there is just one element. For these simple editables we omit the dotted outline and instead only show the solid blue focused node outline.