TASK: Split dialog auto width behaviour from "wide" style
The behaviour of "wide" was changed via https://github.com/neos/neos-ui/pull/4007 to allow contents to be full width.

Now as we no longer position the modal via top: 50% and transform translate -50%
We loose the benefit, that via the positioning we squeeze the content (eg. text or node creation items) into a desired width of 50% - while still allowing the content to be smaller or exceed that size.
To now prevent dialogs to become too large - using `margin: auto` to center to avoid using `transform` which does not work with nested anchored popovers (see advanced link editor options) - we need to reenable that `dialog--wide` restricts its width.

Also the `Dialog` component specified that the `style` is required yet js allows to instantiate the dialog without a style which leads to undefined behaviour and a too large dialog (see keyboard shortcuts, or node variation if there is a lot of text) - to fix that `dialog--wide` becomes the default if nothing is explicitly set.