FEATURE: Show content tree by default when opening the NeosUi for the first time
The toggle state of the content tree (structure tree) is persisted in the local storage.

We as users known to users open it on new projects out of reflex but new users might not even be aware that it exists.

To ensure new users to neos get familiar with the structure of a site its crucial to see the layout of the content nodes which also sometimes eases operations that are not possible in the inline Ui because of too complex visual layouts.