BUGFIX: Prevent clicking structural toolbar to be behind content
In the „below“ case it was possible that the structural toolbar was behind the next content element of the currently selected one. With this change we increase the z-index for the standard and hover case, to also allow selecting the toolbar on hover when the formatting toolbar is also at the bottom.

Resolves: #4054