BUGFIX: Link editor searches through content nodes and shows them in tree
This is rather unintentionally and confusing as these nodes are not selectable and their appearance slows down the search and build up of the tree.
See `06-GetTreeQuery.feature:753` in combination with the new content node.

Additionally, the behat tests now always initialise a content collection to assert that we provide the correct `baseNodeTypeFilter`